v1.4

+++ app/managers/TopicManager.php
@@ -5,38 +5,39 @@ namespace App\Managers;
 use App\Authorizators\VisibilityAuthorizator;
 use App\Constants\TopicMemberRole;
 use App\Core\CacheManager;
+use App\Core\Datetypes\DateTime;
 use App\Entities\TopicEntity;
 use App\Exceptions\AException;
 use App\Exceptions\GeneralException;
+use App\Exceptions\NonExistingEntityException;
 use App\Exceptions\TopicVisibilityException;
 use App\Logger\Logger;
 use App\Repositories\TopicRepository;
+use App\Repositories\TopicRulesRepository;
 use Exception;
 
 class TopicManager extends AManager {
-    private const CACHE_NAMESPACE = 'topics';
-
     private TopicRepository $tr;
-    private CacheManager $cm;
     private TopicMembershipManager $tmm;
     private VisibilityAuthorizator $va;
     private ContentManager $com;
+    private TopicRulesRepository $trr;
 
-    public function __construct(Logger $logger, TopicRepository $topicRepository, TopicMembershipManager $tmm, VisibilityAuthorizator $va, ContentManager $com, EntityManager $entityManager) {
+    public function __construct(Logger $logger, TopicRepository $topicRepository, TopicMembershipManager $tmm, VisibilityAuthorizator $va, ContentManager $com, EntityManager $entityManager, TopicRulesRepository $trr) {
         parent::__construct($logger, $entityManager);
         
         $this->tr = $topicRepository;
         $this->tmm = $tmm;
         $this->va = $va;
-        $this->cm = new CacheManager($logger);
         $this->com = $com;
+        $this->trr = $trr;
     }
 
     public function getTopicById(string $topicId, string $userId) {
         $topic = $this->tr->getTopicById($topicId);
 
         if($topic === null) {
-            throw new TopicVisibilityException('No topic with ID #' . $topicId . ' exists.');
+            throw new NonExistingEntityException('Topic with ID #' . $topicId . ' does not exist.');
         }
 
         $this->checkPrivacy($topic, $userId);
@@ -111,6 +112,10 @@ class TopicManager extends AManager {
     }
 
     public function deleteTopic(string $topicId, string $userId) {
+        if($this->tmm->getFollowRole($topicId, $userId) < TopicMemberRole::OWNER) {
+            throw new GeneralException('You are not authorized to delete this topic.');
+        }
+
         $this->com->deleteTopic($topicId);
 
         $members = $this->tmm->getTopicMembers($topicId, 0, 0, false);
@@ -130,7 +135,7 @@ class TopicManager extends AManager {
         try {
             $this->com->updateTopic($topicId, ['isPrivate' => $isPrivate, 'isVisible' => $isVisible]);
 
-            $this->cm->invalidateCache(self::CACHE_NAMESPACE);
+            $this->cache->invalidateCache(CacheManager::NS_TOPICS);
         } catch(Exception $e) {
             throw $e;
         }
@@ -147,6 +152,8 @@ class TopicManager extends AManager {
 
         try {
             $this->com->pinPost($topicId, $postId);
+
+            $this->cache->invalidateCache(CacheManager::NS_PINNED_POSTS);
         } catch(AException $e) {
             throw $e;
         }
@@ -165,6 +172,8 @@ class TopicManager extends AManager {
 
         try {
             $this->com->pinPost($topicId, $postId, false);
+
+            $this->cache->invalidateCache(CacheManager::NS_PINNED_POSTS);
         } catch(AException $e) {
             throw $e;
         }
@@ -176,6 +185,113 @@ class TopicManager extends AManager {
         $pinnedPostIds = $this->tr->getPinnedPostIdsForTopicId($topicId);
 
         return in_array($postId, $pinnedPostIds);
+    }
+
+    private function getTopicRulesEntityForTopicId(string $topicId) {
+        return $this->trr->getTopicRulesForTopicId($topicId);
+    }
+
+    public function getTopicRulesForTopicId(string $topicId) {
+        $entity = $this->getTopicRulesEntityForTopicId($topicId);
+
+        if($entity === null) {
+            return [];
+        }
+
+        return $entity->getRules();
+    }
+
+    public function hasTopicRules(string $topicId) {
+        $rules = $this->getTopicRulesForTopicId($topicId);
+
+        return !empty($rules);
+    }
+
+    public function addRuleTextToTopicRules(string $topicId, string $ruleText, string $userId) {
+        $entity = $this->getTopicRulesEntityForTopicId($topicId);
+
+        try {
+            if($entity === null) {
+                // create new
+    
+                $rulesetId = $this->trr->createEntityId(EntityManager::TOPIC_RULES);
+    
+                $rulesJson = json_encode([$ruleText]);
+    
+                if(!$this->trr->insertTopicRules($rulesetId, $topicId, $rulesJson, $userId)) {
+                    throw new GeneralException('Could not insert new topic rules.');
+                }
+            } else {
+                // update
+    
+                $rules = $entity->getRules();
+                $rules[] = $ruleText;
+                $rulesJson = json_encode($rules);
+    
+                if(!$this->trr->updateTopicRules($topicId, ['rules' => $rulesJson, 'lastUpdateUserId' => $userId, 'dateUpdated' => DateTime::now()])) {
+                    throw new GeneralException('Could not update topic rules.');
+                }
+            }
+
+            if(!$this->cache->invalidateCache(CacheManager::NS_TOPIC_RULES)) {
+                throw new GeneralException('Could not invalidate cache.');
+            }
+        } catch(AException $e) {
+            throw $e;
+        }
+    }
+
+    public function updateTopicRule(string $topicId, string $userId, int $ruleIndex, string $newText) {
+        $entity = $this->getTopicRulesEntityForTopicId($topicId);
+
+        try {
+            if($entity === null) {
+                throw new GeneralException('Topic has no rules.');
+            }
+
+            $rules = $entity->getRules();
+
+            $rules[$ruleIndex] = $newText;
+
+            $rulesJson = json_encode($rules);
+
+            if(!$this->trr->updateTopicRules($topicId, ['rules' => $rulesJson, 'lastUpdateUserId' => $userId, 'dateUpdated' => DateTime::now()])) {
+                throw new GeneralException('Could not update topic rule.');
+            }
+
+            if(!$this->cache->invalidateCache(CacheManager::NS_TOPIC_RULES)) {
+                throw new GeneralException('Could not invalidate cache.');
+            }
+        } catch(AException $e) {
+            throw $e;
+        }
+    }
+
+    public function deleteTopicRule(string $topicId, string $userId, int $ruleIndex) {
+        $entity = $this->getTopicRulesEntityForTopicId($topicId);
+
+        try {
+            if($entity === null) {
+                throw new GeneralException('Topic has no rules.');
+            }
+
+            $rules = $entity->getRules();
+
+
+            unset($rules[$ruleIndex]);
+
+            $rulesJson = json_encode($rules);
+
+            if(!$this->trr->updateTopicRules($topicId, ['rules' => $rulesJson, 'lastUpdateUserId' => $userId, 'dateUpdated' => DateTime::now()])) {
+                throw new GeneralException('Could not update topic rule.');
+            }
+
+            if(!$this->cache->invalidateCache(CacheManager::NS_TOPIC_RULES)) {
+                throw new GeneralException('Could not invalidate cache.');
+            }
+        } catch(AException $e) {
+            throw $e;
+        }
     }
 }
